% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSNP.R
\docType{data}
\name{dataSNP}
\alias{dataSNP}
\title{SNP dataset without missing values.}
\format{
A text file with 100000 rows and 100 columns.
}
\usage{
dataSNP
}
\description{
100000 simulated SNPs across 100 samples belonging to
two groups with 50 samples each.
}
\keyword{datasets}
