% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_packedancestrymap.R
\name{read_packedancestrymap}
\alias{read_packedancestrymap}
\title{Read Files in PACKEDANCESTRYMAP format}
\usage{
read_packedancestrymap(pref)
}
\arguments{
\item{pref}{The prefix of the file name that contains the genotype data (i.e., without the \code{*.geno}).}
}
\value{
Returns a list containing a single element:
\itemize{
  \item {\code{geno}} {Genotype data as R matrix.}\cr
}
}
\description{
This function loads genotype data in \code{PACKEDANCESTRYMAP} format (binary or compressed).
}
